refactored out common logic
